+++ apps/web/client/tsconfig.json
@@ -18,7 +18,8 @@
     "include": [
         "src",
         "tests",
-        ".next/types/**/*.ts"
+        ".next/types/**/*.ts",
+        "messages/en.d.json.ts"
     ],
     "exclude": [
         "node_modules",
